requirements.txt
mkdocs==1.4.0
mkdocs-awesome-pages-plugin==2.8.0
mkdocs-git-authors-plugin==0.6.4
mkdocs-git-revision-date-localized-plugin==1.0.1
mkdocs-material==8.5.11
mkdocs-material-extensions==1.1.1
mkdocs-mermaid-plugin==0.1.1
mkdocs-minify-plugin==0.5.0
mkdocs-pdf-export-plugin==0.5.10
mkdocs-table-reader-plugin==1.0.0
pymdown-extensions==9.4
markdown-blockdiag==0.8.0
markdown-include==0.6.0
flatten-json==0.1.13
Pygments==2.12.0
mkdocs-enumerate-headings-plugin==0.5.0
mkdocs-blog-plugin==0.25

## Local extras
mkdocs-mermaid2-plugin==0.6.0
mkdocs-with-pdf==0.9.3
mkdocs-plugin-inline-svg==0.1.0
pydyf==0.2.0
cffi==1.15.0
weasyprint==52.5

## The following requirements were added by pip freeze:
actdiag==3.0.0
Babel==2.10.1
beautifulsoup4==4.11.1
blockdiag==3.0.0
bracex==2.2.1
Brotli==1.0.9
click==8.1.3
csscompressor==0.9.5
cssselect2==0.6.0
fonttools==4.33.3
funcparserlib==1.0.0
ghp-import==2.1.0
gitdb==4.0.9
GitPython==3.1.30
html5lib==1.1
htmlmin==0.1.12
importlib-metadata==4.11.3
Jinja2==3.1.2
jsmin==3.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
numpy==1.24.2
nwdiag==3.0.0
packaging==21.3
pandas==1.4.2
Pillow==9.3.0
pycparser==2.21
pyparsing==3.0.8
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyyaml_env_tag==0.1
seqdiag==3.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tinycss2==1.1.1
watchdog==2.1.7
wcmatch==8.3
webcolors==1.11.1
webencodings==0.5.1
zipp==3.8.0
zopfli==0.2.1
